requirements.txt
Django==3.2.8
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pymemcache==3.5.0
requests==2.26.0
gunicorn==20.1.0
